backend/requirements-dev.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
prometheus_client
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio
loguru